add windows stm32 programmer wrapper
